requirements.txt
# ============ CORE FRAMEWORK ============
django>=5.2.0
djangorestframework>=3.14.0

# ============ API DOCUMENTATION ============
drf-spectacular>=0.28.0

# ============ DATABASE ============
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# ============ MACHINE LEARNING ============
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# ============ DATA VALIDATION ============
pydantic>=2.0.0

# ============ LEGACY FASTAPI (opcional) ============
# Mantenemos por compatibilidad pero ya no se usan en producción
fastapi==0.104.1
uvicorn[standard]==0.24.0 